credit-approval-system/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
python-decouple==3.8
django-rq==2.9.0
rq==1.15.1
pandas==2.1.3
openpyxl==3.1.2
celery==5.3.4
python-dateutil==2.8.2